read-cache: make the index write buffer size 2MB to reduce calls into the OS
The Windows OS repo has an index that occupies >260MB. Git was writing
this index <8K at a time, incurring the cost of going to the kernel and
invoking the filesystem very frequently. This change increases the buffer
size to 2MB to give the OS bigger chunks of data to work with each invocation.

This change reduces the cost of write_in_full under ce_write by 3X.

Signed-off-by: Neeraj Singh <neerajsi@ntdev.microsoft.com>